Adjust the login form and signup from styles
