+++ src/common/config.js
@@ -1,0 +1,5 @@
+const appConfig = {
+  strapiURL: process.env.GATSBY_STRAPI_API_URL,
+}
+
+export { appConfig }
