+++ src/locales/ar/messages.json
@@ -18,10 +18,10 @@
   "passPlaceholder": "كلمة السر الخاصة بك",
   "forgotPass": "نسيت كلمة السر؟",
   "or": "أو",
-  "signUsingGithub": " Github تسجيل الدخول باستخدام",
-  "signUsingGoogle": "",
+  "socialSignIn": "{{provider}} تسجيل الدخول باستخدام ",
   "signed": "تم تسجيل الدخول بنجاح",
   "noAccount": "لا تمتلك حساب؟",
+  "hasAccount": "لديك حساب؟",
   "createAccount": "قم بإنشاء حساب",
   "signUp": "إنشاء حساب",
   "name": "الأسم",
