+++ src/locales/en/messages.json
@@ -18,10 +18,10 @@
   "passPlaceholder": "Your Password",
   "forgotPass": "Forgot Password?",
   "or": "Or",
-  "signUsingGithub": "Sign in using Github",
-  "signUsingGoogle": "Sign in using Google",
+  "socialSignIn": "Sign in using {{provider}}",
   "signed": "Signed in Successfuly !",
   "noAccount": "Don't have an account?",
+  "hasAccount": "Have an account already?",
   "createAccount": "Create an Account",
   "signUp": "Sign Up",
   "name": "Name",
